+++ brawlstats/utils.py
@@ -2,6 +2,6 @@ class API:
     def __init__(self, base_url):
         self.base = base_url or 'https://brawlapi.cf/api'
         self.profile = self.base + '/players'
-        self.band = self.base + '/bands'
+        self.club = self.base + '/clubs'
         self.leaderboard = self.base + '/leaderboards'
         self.events = self.base + '/events'